fastapi/requirements.txt
fastapi
pymongo
motor
pydantic
uvicorn
bcrypt
passlib
python-jose[cryptography]
